+++ library/src/test/java/kwasm/format/text/token/FloatLiteralTest.kt
@@ -85,92 +85,92 @@ class FloatLiteralTest {
     @Test
     fun parsesDecimalInts() {
         var actual = FloatLiteral("10")
-        assertThat(actual.value).isWithin(TOLERANCE).of(10.0)
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE).of(10.0)
 
         actual = FloatLiteral("0")
-        assertThat(actual.value).isWithin(TOLERANCE).of(0.0)
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE).of(0.0)
 
         actual = FloatLiteral("1000000000")
-        assertThat(actual.value).isWithin(TOLERANCE).of(1000000000.0)
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE).of(1000000000.0)
     }
 
     @Test
     fun parsesHexInts() {
         var actual = FloatLiteral("0x10")
-        assertThat(actual.value).isWithin(TOLERANCE).of(16.0)
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE).of(16.0)
 
         actual = FloatLiteral("0x0")
-        assertThat(actual.value).isWithin(TOLERANCE).of(0.0)
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE).of(0.0)
 
         actual = FloatLiteral("0xFFFFFFFF")
-        assertThat(actual.value).isWithin(TOLERANCE).of(0xFFFFFFFFL.toDouble())
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE).of(0xFFFFFFFFL.toDouble())
     }
 
     @Test
     fun parsesDecimal_withFraction() {
         var actual = FloatLiteral("1.")
-        assertThat(actual.value).isWithin(TOLERANCE).of(1.0)
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE).of(1.0)
 
         actual = FloatLiteral("1.5")
-        assertThat(actual.value).isWithin(TOLERANCE).of(1.5)
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE).of(1.5)
 
         actual = FloatLiteral("0.123456789")
-        assertThat(actual.value).isWithin(TOLERANCE).of(0.123456789)
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE).of(0.123456789)
 
         actual = FloatLiteral("123456789.123456789")
-        assertThat(actual.value).isWithin(TOLERANCE).of(123456789.123456789)
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE).of(123456789.123456789)
     }
 
     @Test
     fun parsesHex_withFraction() {
         var actual = FloatLiteral("0xF.")
-        assertThat(actual.value).isWithin(TOLERANCE).of(15.0)
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE).of(15.0)
 
         actual = FloatLiteral("0x10.a")
-        assertThat(actual.value).isWithin(TOLERANCE).of(16.0 + 10 / 16.0)
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE).of(16.0 + 10 / 16.0)
     }
 
     @Test
     fun parsesDecimal_withExponent() {
         var actual = FloatLiteral("1e10")
-        assertThat(actual.value).isWithin(TOLERANCE).of(1e10)
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE).of(1e10)
 
         actual = FloatLiteral("3e-5")
-        assertThat(actual.value).isWithin(TOLERANCE).of(3e-5)
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE).of(3e-5)
     }
 
     @Test
     fun parsesHex_withExponent() {
         var actual = FloatLiteral("0xFp2")
-        assertThat(actual.value).isWithin(TOLERANCE).of(15 * 2.0.pow(2))
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE).of(15 * 2.0.pow(2))
 
         actual = FloatLiteral("0xABp-2")
-        assertThat(actual.value).isWithin(TOLERANCE).of(0xAB * 2.0.pow(-2))
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE).of(0xAB * 2.0.pow(-2))
     }
 
     @Test
     fun parsesDecimal_withFraction_andExponent() {
         var actual = FloatLiteral("1.e10")
-        assertThat(actual.value).isWithin(TOLERANCE).of(1e10)
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE).of(1e10)
 
         actual = FloatLiteral("1.5e10")
-        assertThat(actual.value).isWithin(TOLERANCE).of(1.5e10)
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE).of(1.5e10)
 
         actual = FloatLiteral("3.12345e-4")
-        assertThat(actual.value).isWithin(TOLERANCE).of(3.12345e-4)
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE).of(3.12345e-4)
     }
 
     @Test
     fun parsesHex_withFraction_andExponent() {
         var actual = FloatLiteral("0xF.p2")
-        assertThat(actual.value).isWithin(TOLERANCE).of(15 * 2.0.pow(2))
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE).of(15 * 2.0.pow(2))
 
         actual = FloatLiteral("0xF.ap2")
-        assertThat(actual.value).isWithin(TOLERANCE)
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE)
             .of((15 + 10 / 16.0) * 2.0.pow(2))
 
         actual = FloatLiteral("0xAB.1p-2")
-        assertThat(actual.value).isWithin(TOLERANCE)
+        assertThat(actual.value.toDouble()).isWithin(TOLERANCE)
             .of((0xAB + 1 / 16.0) * 2.0.pow(-2))
     }
